infra/control-plane/app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
boto3==1.36.7
python-jose[cryptography]==3.3.0
pydantic==2.10.5
httpx==0.28.1
